Provides group affinity partitioning capability

+++ pkg/controllers/job/job_controller_handler.go
@@ -175,18 +175,26 @@ func (cc *jobcontroller) addPod(obj interface{}) {
 		return
 	}
 
+	taskName, found := pod.Annotations[batch.TaskSpecKey]
+	if !found {
+		klog.Infof("Failed to find taskName of Pod <%s/%s>, skipping",
+			pod.Namespace, pod.Name)
+		return
+	}
+
 	if pod.DeletionTimestamp != nil {
 		cc.deletePod(pod)
 		return
 	}
 
 	req := apis.Request{
-		Namespace: pod.Namespace,
-		JobName:   jobName,
-		JobUid:    jobUid,
-		PodName:   pod.Name,
-		PodUID:    pod.UID,
-
+		Namespace:  pod.Namespace,
+		JobName:    jobName,
+		JobUid:     jobUid,
+		PodName:    pod.Name,
+		PodUID:     pod.UID,
+		TaskName:   taskName,
+		Partition:  getPodPartition(pod),
 		Event:      bus.PodPendingEvent,
 		JobVersion: int32(dVersion),
 	}
@@ -298,13 +306,13 @@ func (cc *jobcontroller) updatePod(oldObj, newObj interface{}) {
 	}
 
 	req := apis.Request{
-		Namespace: newPod.Namespace,
-		JobName:   jobName,
-		JobUid:    jobUid,
-		TaskName:  taskName,
-		PodName:   newPod.Name,
-		PodUID:    newPod.UID,
-
+		Namespace:  newPod.Namespace,
+		JobName:    jobName,
+		JobUid:     jobUid,
+		TaskName:   taskName,
+		PodName:    newPod.Name,
+		PodUID:     newPod.UID,
+		Partition:  getPodPartition(newPod),
 		Event:      event,
 		ExitCode:   exitCode,
 		JobVersion: int32(dVersion),
@@ -368,13 +376,13 @@ func (cc *jobcontroller) deletePod(obj interface{}) {
 	}
 
 	req := apis.Request{
-		Namespace: pod.Namespace,
-		JobName:   jobName,
-		JobUid:    jobUid,
-		TaskName:  taskName,
-		PodName:   pod.Name,
-		PodUID:    pod.UID,
-
+		Namespace:  pod.Namespace,
+		JobName:    jobName,
+		JobUid:     jobUid,
+		TaskName:   taskName,
+		PodName:    pod.Name,
+		PodUID:     pod.UID,
+		Partition:  getPodPartition(pod),
 		Event:      bus.PodEvictedEvent,
 		JobVersion: int32(dVersion),
 	}
